Better display of general parameters in REST interface
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3317 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/samson/worker/SamsonWorker.h
@@ -192,7 +192,7 @@ namespace samson {
             output << "[";
             for ( int i = 0 ; i < n ; i++ )
             {
-                output << "[" << (i-n+1) << "," << values[i] << "]";
+			   output << "[" << (i-n+1) << "," << ((int)values[i]) << "]";
                 if( i != (n-1) )
                     output << ",";
             }
